split file utils into zran_file_util
